zadanie2/requirements.txt
flask
celery
redis
sqlalchemy
pydantic
wtforms
httpx
pytest
flask_sqlalchemy
flask_wtf
python-dotenv
email_validator
uwsgi
hypercorn
